P1_MiniLM/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
requests
python-dotenv
transformers
torch
pandas
python-multipart
streamlit
